+++ src/afterglow/show.clj
@@ -919,97 +919,32 @@
       [cue active])))
 
 (defn add-midi-control-to-cue-mapping
-  "Cause the specified cue from the [[*show*]] cue grid to be
-  triggered by receipt of the specified note (when `kind` is `:note`)
-  or controller-change (when `kind` is `:control`) message with a
-  non-zero velocity or control value. This allows generic MIDI
-  controllers, which do not have enough pads or feedback capabilities
-  to act as a full grid controller like the Ableton Push, to still
-  provide a physical means of triggering cues. The desired cue is
-  identified by passing in its `x` and `y` coordinates within the show
-  cue grid.
-
-  Afterglow will attempt to provide feedback about the progress of the
-  cue by sending note on/off or control-change values to the same
-  controller when the cue starts and ends. The note velocities or
-  control values used can be changed by passing in different values
-  with `:feedback-on` and `:feedback-off`, and this behavior can be
-  suppressed entirely by passing `false` with `:feedback-on`.
-
-  If the controller is pressure-sensitive and you would like to have
-  the velocity information passed on to any cue variables which are
-  configured to respond to it, pass `true` with `use-velocity?`. Doing
-  this also enables responsiveness to aftertouch messages, which can
-  adjust the cue variable for as long as you are holding down the key
-  or pad, if your controller sends them (and `kind` is `:note`).
-
-  Afterglow assumes the control is momentary, meaning it sends a note
-  off (or control value of 0) as soon as it is released, and a second
-  press will be used to end the cue unless the cue uses the `:held`
-  modifier to indicate it should be ended when the button is released.
-  If your controller does not have momentary buttons and already
-  requires a second press to turn off the note or control value, pass
-  `false` with `:momentary?` and Afterglow will always end cues when
-  it receives a control value of 0, even if cues are not marked as
-  `:held`."
+  "This function was moved
+  to [[afterglow.effects.cues/add-midi-control-to-cue-mapping]] in
+  version 0.1.6 in order to resolve circular dependency issues
+  introduced when adding velocity and aftertouch support. This stub
+  was left behind for backwards compatibility, but will be removed in
+  the next version, so please update your code to find it in its new
+  location."
+  {:deprecated "0.1.6"}
   [midi-device-name channel kind note x y & {:keys [feedback-on feedback-off use-velocity? momentary?]
                                              :or {feedback-on 127 feedback-off 0 momentary? true}}]
-  {:pre [(some? *show*) (#{:control :note} kind) (some? midi-device-name) (integer? channel) (<= 0 channel 15)
-         (integer? note) (<= 0 note 127) (integer? x) (<= 0 x) (integer? y) (<= 0 y)
-         (or (not feedback-on) (and (integer? feedback-on) (<= 0 feedback-on 127)))
-         (integer? feedback-off) (<= 0 feedback-off 127)]}
-  (let [show *show*  ; Bind so we can pass it to update functions running on another thread
-        feedback-device (when feedback-on (midi/find-midi-out midi-device-name))
-        our-id (atom nil)  ; Track when we have created an effect
-        midi-handler (fn [msg]
-                       (with-show show
-                         ;; See if the cue exists and is running
-                         (let [[cue active] (find-cue-grid-active-effect show x y)
-                               velocity (:velocity msg)]
-                           (when cue
-                             (if (and (pos? velocity) (not= (:command msg) :note-off))
-                               ;; Control or note has been pressed
-                               (if (and active (not (:held cue)))
-                                 (end-effect! (:key cue))
-                                 (let [vars (when use-velocity? (controllers/starting-vars-for-velocity cue velocity))]
-                                   (reset! our-id (add-effect-from-cue-grid! x y :var-overrides vars))))
-                               ;; Control has been released
-                               (when (and (some? @our-id) (or (:held cue) (not momentary?)))
-                                 (end-effect! (:key cue) :when-id @our-id)
-                                 (reset! our-id nil)))))))]
-    (when feedback-device  ; Set up to give feedback as cue activation changes
-      (controllers/add-cue-feedback! (:cue-grid show) x y feedback-device channel kind note
-                                     :on feedback-on :off feedback-off)
-      (let [[cue active] (find-cue-grid-active-effect show x y)]  ; Was already active, so reflect that
-        (when active (case kind
-                      :control (overtone.midi/midi-control feedback-device note feedback-on channel)
-                      :note (overtone.midi/midi-note-on feedback-device note feedback-on channel)))))
-    ;; TODO: Add aftertouch support
-    (case kind
-      :control (midi/add-control-mapping midi-device-name channel note (str "show:" (:id show) ":cue-control-" x "-" y)
-                                         midi-handler)
-      :note (midi/add-note-mapping midi-device-name channel note (str "show:" (:id show) ":cue-note-" x "-" y)
-                                   midi-handler))))
+  (require '[afterglow.effects.cues])
+  ((resolve 'afterglow.effects.cues/add-midi-control-to-cue-mapping) midi-device-name channel kind note x y
+   :feedback-on feedback-on :feedback-off feedback-off :use-velocity? use-velocity? :momentary? momentary?))
 
 (defn remove-midi-control-to-cue-mapping
-  "Stop triggering the specified cue from the [[*show*]] cue grid upon
-  receipt of the specified note or controller-change message. The
-  desired cue is identified by passing in its `x` and `y` coordinates
-  within the show cue grid."
+"This function was moved
+  to [[afterglow.effects.cues/remove-midi-control-to-cue-mapping]] in
+  version 0.1.6 in order to resolve circular dependency issues
+  introduced when adding velocity and aftertouch support. This stub
+  was left behind for backwards compatibility, but will be removed in
+  the next version, so please update your code to find it in its new
+  location."
+  {:deprecated "0.1.6"}
   [midi-device-name channel kind note x y]
-  {:pre [(some? *show*) (#{:control :note} kind) (some? midi-device-name) (integer? channel) (<= 0 channel 15)
-         (integer? note) (<= 0 note 127) (integer? x) (<= 0 x) (integer? y) (<= 0 y)]}
-  (let [feedback-device (midi/find-midi-out midi-device-name)
-        feedback (controllers/clear-cue-feedback! (:cue-grid *show*) x y feedback-device channel kind note)]
-    (when feedback  ; We had been giving feedback, see if we need to turn it off
-      (let [[cue active] (find-cue-grid-active-effect *show* x y)]
-        (when active (case kind
-                       :control (overtone.midi/midi-control feedback-device note (second feedback) channel)
-                       :note (overtone.midi/midi-note-on feedback-device note (second feedback) channel))))))
-  (case kind
-    :control (midi/remove-control-mapping midi-device-name channel note
-                                          (str "show:" (:id *show*) ":cue-control-" x "-" y))
-    :note (midi/remove-note-mapping midi-device-name channel note (str "show:" (:id *show*) ":cue-note-" x "-" y))))
+  (require '[afterglow.effects.cues])
+  ((resolve 'afterglow.effects.cues/remove-midi-control-to-cue-mapping) midi-device-name channel kind note x y))
 
 (defn- address-map-internal
   "Helper function which returns a sorted map whose keys are all